fix: fix translation of unscheduled in German
